Move chartpage
