+++ src/pages/graphData/components/chart/i18n/de.ts
