+++ src/pages/graphData/components/chart/i18n/en.ts
