+++ src/pages/graphData/components/chart/i18n/index.ts
