+++ src/pages/graphData/components/chart/scripts/models/chartDataFunctions/FuelBurnt.ts
