+++ src/pages/graphData/components/chart/scripts/models/chartDataFunctions/FuelConsumption.ts
