+++ src/pages/graphData/components/chart/scripts/models/chartDataFunctions/FuelPricing.ts
