+++ src/pages/graphData/components/chart/scripts/models/chartDataFunctions/MoneySpent.ts
