+++ src/pages/graphData/components/chart/scripts/models/chartDataFunctions/index.ts
