+++ src/pages/graphData/components/chart/scripts/models/registerChartData.ts
